Move notification objects into their own module
